requirements.txt
discord.py
requests
PythonGists